refactor(core): hoist block metadata resolver out of API crates
