+++ crates/minibf/src/lib.rs
@@ -27,7 +27,6 @@ use dolos_core::{
     TemporalKey, TxOrder,
 };
 
-pub mod block_meta;
 mod cache;
 mod error;
 pub(crate) mod hacks;
